Make integration test viable including auth middleware
